collection count, with refactor
